Edits for evaluator

+++ src/main/java/Jalgo/strategy/Strategy.java
@@ -6,6 +6,6 @@ public interface Strategy {
      * Core component. Algorithm to decide buy, sell, hold is in this method.
      * @param float current price that stock sold at
      */
-    public DecisionEnum evaluate (float price);
+    public DecisionEnum evaluate (double price);
 
 }
